fix(survival): compute the random survival forest log-rank variance in double
CodeQL cs/loss-of-precision (alerts 15348-15352), RandomSurvivalForest.cs:326-327: the
log-rank variance numerator and denominator were int products that overflow once about
1,291 subjects are at risk, corrupting split selection.

+++ src/SurvivalAnalysis/RandomSurvivalForest.cs
@@ -323,8 +323,8 @@ public partial class RandomSurvivalForest<T> : SurvivalModelBase<T>
             // Variance contribution
             if (totalAtRisk > 1)
             {
-                variance += (double)(leftAtRisk * rightAtRisk * totalEvents * (totalAtRisk - totalEvents))
-                           / (totalAtRisk * totalAtRisk * (totalAtRisk - 1));
+                variance += (double)leftAtRisk * rightAtRisk * totalEvents * (totalAtRisk - totalEvents)
+                           / ((double)totalAtRisk * totalAtRisk * (totalAtRisk - 1));
             }
         }
